search completed

+++ server/Utills/Decompress/decompression.js
@@ -5,7 +5,7 @@ const deleteDir = require("del");
 const util = require("util");
 
 const { parsing } = require("../Parse/parser");
-const { createNode, insert } = require("../TrieDs/trieNode");
+const { createNode, insert,search } = require("../TrieDs/trieNode");
 const { subDecompress } = require("./subDecompress");
 const root = new createNode("");
 
@@ -16,24 +16,27 @@ const decompression = async (dataFolder, distFolder) => {
     const folders = fs.readdirSync(dataFolder);
 
     for (const folder of folders) {
-      xmlFilePath = await subDecompress(dataFolder, folder, distFolder);
-      console.log(xmlFilePath);
-      //Parsing
-      for (const xmlFile of xmlFilePath) {
-          let match = await parsing(distFolder, xmlFile.path);
+      if(search(folder.split(".")[0],root)=="Not Found"){
 
-          if (match == folder.split(".")[0]) {
-            //trie
-        insert(
-          folder.split(".")[0],
-          root,
-          `./${distFolder}/${xmlFile.finalPath}`
-        );
-          }
+        xmlFilePath = await subDecompress(dataFolder, folder, distFolder);
+        console.log(xmlFilePath);
+        //Parsing
+        for (const xmlFile of xmlFilePath) {
+            let match = await parsing(distFolder, xmlFile.path);
+  
+            if (match == folder.split(".")[0]) {
+              //trie
+          insert(
+            folder.split(".")[0],
+            root,
+            `${xmlFile.finalPath}`
+          );
+            }
+        }
+        xmlFilePath = [];
+        //delete dir
+        await deleteDir(`./${distFolder}`)
       }
-      xmlFilePath = [];
-      delete dir
-      await deleteDir(`./${distFolder}`)
     }
     console.log(util.inspect(root, false, null, true /* enable colors */));
   } catch (error) {
